Fix warnings for wrong type to %.*s format
Field widths must be of type int.

In file included from testaes.c:7:0:
./ext/cutest.h: In function ‘test_do_run__’:
./ext/cutest.h:302:13: error: field precision specifier ‘.*’ expects
argument of type ‘int’, but argument 2 has type ‘long unsigned int’
[-Werror=format=]
             printf("%.*s", sizeof(spaces) - n, spaces);

+++ include/cutest.h
@@ -300,7 +300,7 @@ test_do_run__(const struct test__* test)
         n = test_print_in_color(CUTEST_COLOR_DEFAULT_INTENSIVE__, "Test %s... ", test->name);
         memset(spaces, ' ', sizeof(spaces));
         if(n < sizeof(spaces))
-            printf("%.*s", sizeof(spaces) - n, spaces);
+            printf("%.*s", (int) (sizeof(spaces) - n), spaces);
     } else {
         test_current_already_logged__ = 1;
     }
